物理部分の警告をなくした

+++ Adollib/Adollib/Scripts/Main/Adollib.cpp
@@ -10,11 +10,11 @@ namespace Adollib {
 	float Al_Global::second_per_frame = 0;
 	float Al_Global::second_per_game = 0;
 
-	float Al_Global::base_Coefficient_of_friction = 0.6;
-	float Al_Global::base_resituation = 0.2;
+	float Al_Global::base_Coefficient_of_friction = 0.6f;
+	float Al_Global::base_resituation = 0.2f;
 
 	float get_P_by_f(float M, float F) {
-		int i = M / F;
+		int i = (int)(M / F);
 		return M - i * F;
 	};
